Auth

+++ src/main/java/com/example/dev_diaries/services/NotesService.java
@@ -8,29 +8,43 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import javax.management.RuntimeErrorException;
+
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import com.example.dev_diaries.models.Format;
 import com.example.dev_diaries.models.Note;
 import com.example.dev_diaries.models.Tag;
+import com.example.dev_diaries.models.User;
 import com.example.dev_diaries.repositories.NotesRepository;
 import com.example.dev_diaries.repositories.TagsRepository;
+import com.example.dev_diaries.repositories.UserRepository;
 import com.example.dev_diaries.specifications.NotesSpecification;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class NotesService {
     TaggingService taggingService;
     NotesRepository notesRepository;
     TagsRepository tagsRepository;
+    UserRepository userRepository;
 
-    public NotesService(TaggingService taggingService, NotesRepository notesRepository, TagsRepository tagsRepository) {
+    public NotesService(TaggingService taggingService, NotesRepository notesRepository, TagsRepository tagsRepository, UserRepository userRepository) {
         this.taggingService = taggingService;
         this.notesRepository = notesRepository;
         this.tagsRepository = tagsRepository;
+        this.userRepository = userRepository;
     }
 
-    public Note createNote(Note note) {
+    private User getUser(String email){
+        return userRepository.findByEmail(email).orElseThrow(
+            () -> new EntityNotFoundException("User not found")
+        );
+    }
+    public Note createNote(Note note, String userEmail) {
+        User user = getUser(userEmail);
         String text = Stream.of(note.getContent(), note.getTitle())
                 .filter(Objects::nonNull)
                 .collect(Collectors.joining(" "));
@@ -41,19 +55,17 @@ public class NotesService {
                 .orElseGet(() -> tagsRepository.save(new Tag(tagname)))).collect(Collectors.toSet());
 
         note.setTags(tags);
+        note.setUser(user);
         return notesRepository.save(note);
     }
 
-    public List<Note> getAllNotes() {
-        return notesRepository.findAll();
-    }
-
-    public Optional<Note> getNoteBYId(UUID id) {
-        return notesRepository.findById(id);
-    }
-
-    public Note updateNote(UUID id, Note newNote) {
+    public Note updateNote(UUID id, Note newNote, String userEmail) {
+        User user = getUser(userEmail);
         Note oldNote = notesRepository.findById(id).orElseThrow();
+
+        if(!oldNote.getUser().getId().equals(user.getId())){
+            throw new RuntimeException("Unauthorized: you do not own this note");
+        }
 
         if (newNote.getTitle() != null && !newNote.getTitle().isBlank())
             oldNote.setTitle(newNote.getTitle());
@@ -75,8 +87,9 @@ public class NotesService {
         return notesRepository.save(oldNote);
     }
 
-    public List<Note> searchNotes(String keyword, String tagName, Format format) {
-        Specification<Note> spec = Specification.where(Specification.unrestricted());
+    public List<Note> searchNotes(String keyword, String tagName, Format format, String userEmail) {
+        User user = getUser(userEmail);
+        Specification<Note> spec = Specification.where(NotesSpecification.belongsToUser(user));
         if(keyword != null && !keyword.isBlank()){
             spec = spec.and(NotesSpecification.containsKeyword(keyword));
         }
@@ -92,7 +105,12 @@ public class NotesService {
         return notesRepository.findAll(spec);
     }
 
-    public void deleteNote(UUID id) {
+    public void deleteNote(UUID id, String userEmail) {
+        User user = getUser(userEmail);
+        Note note = notesRepository.findById(id).orElseThrow();
+        if(!note.getUser().getId().equals(user.getId())){
+            throw new RuntimeException("Unauthorized");
+        }
         notesRepository.deleteById(id);
     }
 }
